+++ .travis.yml
@@ -1,129 +1,28 @@
-dist:     trusty    # используем Ubuntu 14.04 Trusty Tahr (а не 12.02 Precise Pangolin)
-sudo:     required  # используем Virtual Machine (а не Docker container)
+dist: bionic
+
+sudo: required
 
 language: c
 
-#
-# Build Dependencies
-#
+os: linux
+
+compiler: gcc
+  
 before_script:
+  - sudo apt-get -y update
+  - sudo apt-get install gcc-7 g++-7 cmake
 
-  #
-  # Build Folder
-  #
-  - mkdir build
-  - cd build
-  #
-  # Build gtest via cmake
-  #
-  - |
-    if [ ${TASK} == "cmake_test" ]; then
-    set -e
-    wget -nc https://github.com/google/googletest/archive/release-1.7.0.zip
-    unzip -n release-1.7.0.zip
-    mv googletest-release-1.7.0 gtest && cd gtest
-    cmake . && make
-    mkdir lib && mv libgtest.a lib
-    cd ..
-    rm -rf release-1.7.0.zip
-    rm -rf build
-    mkdir build && cd build
-    PLUGINS="-DPLUGIN_LZ4=ON -DPLUGIN_DENSE_PARSER=ON"
-    cmake .. -DGOOGLE_TEST=ON -DGTEST_ROOT=$PWD/../gtest/ ${PLUGINS}
-    make
-    cd ..
-    ./testxgboost
-    rm -rf build
-    fi
-
-
-  #
-  # Update / Install CMake
-  #
-  - |
-    if [[ "${TRAVIS_OS_NAME}" == "linux" ]]; then
-      mkdir -p external/cmake
-      pushd external/cmake
-      wget https://cmake.org/files/v3.8/cmake-3.8.0-Linux-x86_64.sh
-      chmod +x cmake-*-Linux-x86_64.sh
-      ./cmake-*-Linux-x86_64.sh --exclude-subdir --skip-license
-      export PATH="${PWD}/bin:$PATH"
-      popd
-    else
-      if ! brew ls --version cmake &>/dev/null; then brew update; brew install cmake; fi
-    fi
-
-#
-# Build Matrix
-#
-matrix:
-  include:
-
-  #
-  # CppCheck
-  #
-  - os: linux
-    env:
-      - TEST="CppCheck"
-    addons:
-      apt:
-        sources:
-          - ubuntu-toolchain-r-test
-        packages:
-          - gcc-6
-          - g++-6
-    script:
-      - cmake -DENABLE_CPPCHECK=ON -DCMAKE_CXX_COMPILER="g++-6" ..
-      - make
-      - make check      
-
-  #
-  # Valgrind
-  #
-  - os: linux
-    env:
-      - TEST="Valgrind"
-    addons:
-      apt:
-        sources:
-          - ubuntu-toolchain-r-test
-        packages:
-          - gcc-6
-          - g++-6
-          - valgrind
-    script:
-      - cmake -DCMAKE_CXX_COMPILER="g++-6" ..
-      - make
-      - ctest -T memcheck
-
-  #
-  # OS configurations check
-  #
-  - os: linux
-    compiler:
-      - gcc
-      - clang
-    env:
-    # список переменных окружения, влияющих на матрицу сборки:
-      - TARGET_CPU=amd64 BUILD_CONFIGURATION=Debug
-      - TARGET_CPU=amd64 BUILD_CONFIGURATION=Release
-      - TARGET_CPU=x86 BUILD_CONFIGURATION=Debug
-      - TARGET_CPU=x86 BUILD_CONFIGURATION=Release
-    script:
-    # скрипт сборки и тестирования проекта:
-      - cmake ..
-      - make
-      - ctest --output-on-failure
-
-deploy:
-    # выкладываем tagged-коммиты на GitHub Releases:
-    provider:     releases
-    #file:         <package-file>
-    #skip_cleanup: true
-    #overwrite:    true
-
-    api_key:
-        secure:   bgNABu38bk4n5KSCXkCpsxy8RldbwCgTlsxIk5IB7/I9rJSyoDqMRZsbtP32vDyKtuDhf2gFoeXS4n34/Psik5gsxCc9R2cwYDBCmzU3viw7WSdTP74NPB5XsWSHAusvru8LVpQgyP+nVyO1FPFpvauqmcNZscibsl7xqHOz6cEN9TOg99cH4A4SGtaA4+IBhc7jcHxh9OnjJATiBQc7Hs4TKINGlBU5ozBlKymG+Qoferc9iYT58kVvNDp/X4WUiW1D8JFnK4BKgT70ZhaawhS7UC/qPj3sPDP4+OszI9GHJqA82qY3sqRMkele2luQaDVqJpoJQNr58QHE4GfNfJs6uNwpVAWOvjPauLVvfI7EiJcENf1pH0GZZe8rPoi0QnddJSRjJzKVgpGz4rbPfpaPGArxtTCFMqCcD+x2E8y/hoT3Ajqz/NViLO3FAfhXrfzbaT62J5CWZ8gDHYFYsGrx+monLOxFmy+KBxIGnS1jK3E2lQuPfO93CkLpMFOThwLXIv8qEohdmPXdDy5EqrKcZGQg5MxaTh8tiNNxXDm8pBtec0Nwz5FbYWSQ2CZ+fj8NCeZHHqh4IznB8DfxteEEIF2qUAqNOtqph3ytJcu7U3wnuOhNrZNH9XwiW24D53AP2kB7uFVBSj2eHQWcKLvQTjYSwR5Xh7NSrPvwApY=
-
-    on:
-        tags:     true
+script:
+  - sudo apt-get install -y libgtest-dev
+  - cd /usr/src/gtest
+  - sudo cmake CMakeLists.txt && sudo make
+  - sudo cp *.a /usr/lib
+  - cd "${TRAVIS_BUILD_DIR}"
+  - mkdir build && cd build
+  - cmake .. && make
+#  test stage
+  - sudo apt-get install -y valgrind cppcheck gcovr lcov
+  - cppcheck --inconclusive --suppress=missingIncludeSystem --check-config --enable=all --language=c ../main.c ../compilation_builder.h ../compilation_builder.c 
+  - valgrind --leak-check=full --leak-resolution=med --vgdb=no --track-origins=yes ./main 100
+  - ctest --output-on-failure
+  - gcovr -r ..
