+++ CMakeLists.txt
@@ -1,137 +1,32 @@
-cmake_minimum_required(VERSION 3.6) # Проверка версии CMake.
-									# Если версия установленой программы
-									# старее указаной, произайдёт аварийный выход.
-project(main)
 
-include(ExternalProject)
-include_directories(BEFORE ${GTEST_INCLUDE_DIRS})
-find_package(Git REQUIRED)
+cmake_minimum_required(VERSION 2.6)
+project(tests)
 
-set_source_files_properties(src/http_parser/http_parser.c PROPERTIES LANGUAGE CXX)
-set(CMAKE_CXX_STANDARD 14)
-set(CMAKE_EXPORT_COMPILE_COMMANDS ON)
+set(CMAKE_C_STANDARD 99)
+#set(CMAKE_CXX_STANDARD 11)
+#set(SMAKE_CXX_STANDARD_REQUIRED ON)
+#set(CMAKE_CXX_FLAGS "-g -O0 -Wall -fprofile-arcs -ftest-coverage")
+set(CMAKE_C_FLAGS "-g -O0 -Wall -Werror -Wpedantic -fprofile-arcs -ftest-coverage")
+# -ftest-coverage - Create data files needed by gcov
+# -fprofile-arcs - Insert arc-based program profiling code 
 
-# ------------------------------------------------------------------------------
-# system configuration choice
-# ------------------------------------------------------------------------------
+set(CMAKE_EXE_LINKER_FLAGS "-lgcov --coverage")
 
-if ("${TARGET_CPU}" STREQUAL "amd64")
-    set (CMAKE_SIZEOF_VOID_P 8)
+set(LIB_SOURCE_FILE functionality.h functionality.c)
 
-    set_property (GLOBAL PROPERTY FIND_LIBRARY_USE_LIB64_PATHS TRUE)
-    set_property (GLOBAL PROPERTY FIND_LIBRARY_USE_LIB32_PATHS FALSE)
-elseif ("${TARGET_CPU}" STREQUAL "x86")
-    set (CMAKE_SIZEOF_VOID_P 4)
+add_library(_all_headers STATIC ${LIB_SOURCE_FILE})
 
-    set_property (GLOBAL PROPERTY FIND_LIBRARY_USE_LIB64_PATHS FALSE)
-    set_property (GLOBAL PROPERTY FIND_LIBRARY_USE_LIB32_PATHS TRUE)
+find_package(GTest REQUIRED)
+include_directories(${GTEST_INCLUDE_DIRS})
 
-    if (GCC)
-        set (CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -m32")
-        set (CMAKE_C_FLAGS   "${CMAKE_C_FLAGS} -m32")
-    endif ()
-#else ()
-#   message (FATAL_ERROR "Unsupported CPU: ${TARGET_CPU}")
-endif ()
+add_executable(tests test.cpp)
+add_executable(main main.c)
 
-# ------------------------------------------------------------------------------
-# CppCheck
-# ------------------------------------------------------------------------------
-
-if(ENABLE_CPPCHECK)
-
-    list(APPEND CPPCHECK_CMAKE_ARGS
-        "-DCMAKE_INSTALL_PREFIX=${CMAKE_BINARY_DIR}"
-    )
-
-    ExternalProject_Add(
-        cppcheck
-        GIT_REPOSITORY      https://github.com/danmar/cppcheck.git
-        GIT_TAG             1.79
-        GIT_SHALLOW         1
-        CMAKE_ARGS          ${CPPCHECK_CMAKE_ARGS}
-        PREFIX              ${CMAKE_BINARY_DIR}/external/cppcheck/prefix
-        TMP_DIR             ${CMAKE_BINARY_DIR}/external/cppcheck/tmp
-        STAMP_DIR           ${CMAKE_BINARY_DIR}/external/cppcheck/stamp
-        DOWNLOAD_DIR        ${CMAKE_BINARY_DIR}/external/cppcheck/download
-        SOURCE_DIR          ${CMAKE_BINARY_DIR}/external/cppcheck/src
-        BINARY_DIR          ${CMAKE_BINARY_DIR}/external/cppcheck/build
-    )
-
-    list(APPEND CPPCHECK_ARGS
-        --enable=warning,style,performance,portability,unusedFunction
-        --std=c++11
-        --verbose
-        --error-exitcode=1
-        --language=c++
-        -DMAIN=main
-        -I ${CMAKE_SOURCE_DIR}/include
-        ${CMAKE_SOURCE_DIR}/include/*.h
-        ${CMAKE_SOURCE_DIR}/src/*.cpp
-        ${CMAKE_SOURCE_DIR}/test/*.cpp
-    )
-
-    add_custom_target(
-        check
-        COMMAND ${CMAKE_BINARY_DIR}/bin/cppcheck ${CPPCHECK_ARGS}
-        COMMENT "running cppcheck"
-    )
-
-endif()
-
-# ------------------------------------------------------------------------------
-# Valgrind
-# ------------------------------------------------------------------------------
-
-set(MEMORYCHECK_COMMAND_OPTIONS "${MEMORYCHECK_COMMAND_OPTIONS} --leak-check=full")
-set(MEMORYCHECK_COMMAND_OPTIONS "${MEMORYCHECK_COMMAND_OPTIONS} --track-fds=yes")
-set(MEMORYCHECK_COMMAND_OPTIONS "${MEMORYCHECK_COMMAND_OPTIONS} --trace-children=yes")
-set(MEMORYCHECK_COMMAND_OPTIONS "${MEMORYCHECK_COMMAND_OPTIONS} --error-exitcode=1")
-
-# ------------------------------------------------------------------------------
-# gtest definitions
-# ------------------------------------------------------------------------------
-
-configure_file(CMakeLists.txt.in
-        googletest-download/CMakeLists.txt)
-execute_process(COMMAND ${CMAKE_COMMAND} -G "${CMAKE_GENERATOR}" .
-        WORKING_DIRECTORY ${CMAKE_BINARY_DIR}/googletest-download )
-execute_process(COMMAND ${CMAKE_COMMAND} --build .
-        WORKING_DIRECTORY ${CMAKE_BINARY_DIR}/googletest-download )
-
-add_subdirectory(${CMAKE_BINARY_DIR}/googletest-src
-        ${CMAKE_BINARY_DIR}/googletest-build)
+target_link_libraries(tests ${GTEST_LIBRARIES} gtest_main pthread)
+target_link_libraries(tests _all_headers)
+target_link_libraries(main _all_headers)
+target_link_libraries(tests -fprofile-arcs)
 
 enable_testing()
-add_subdirectory(test)
 
-# ------------------------------------------------------------------------------
-# Main Executable
-# ------------------------------------------------------------------------------
-
-include_directories("./include")
-
-file(GLOB sources 
-    "${PROJECT_SOURCE_DIR}/include/hw1/*.h"
-    "${PROJECT_SOURCE_DIR}/src/*.c")
-
-add_executable(main ${sources})		# Создает исполняемый файл с именем main
-									# из исходника main.cpp
-
-set_property(TARGET main PROPERTY C_STANDARD 99)
-
-# ------------------------------------------------------------------------------
-# Tests
-# ------------------------------------------------------------------------------
-
-include(CTest)
-
-macro(do_test str)
-    add_executable(test_${str} test/${str}.cpp)
-    add_test(test_${str} test_${str})
-    target_compile_definitions(test_${str} PRIVATE VIRTUAL=virtual)
-endmacro(do_test)
-
-do_test(err)
-do_test(main)
-do_test(struct)
+add_test(NAME tests COMMAND tests)
