--- include/hw1/io.h
@@ -1,8 +1,0 @@
-#ifndef _IO_H_
-#define _IO_H_
-
-array_return input( array_return );
-
-array_return output( array_return );
-
-#endif
