--- include/hw1/numfreq.h
@@ -1,6 +1,0 @@
-#ifndef _NUMFREQ_H_
-#define _NUMFREQ_H_
-
-array_return number_frequency( array_return );
-
-#endif
